+++ packages/react-templates/src/index.ts
@@ -7,6 +7,8 @@ import NotificationFeed, {NotificationFeedProps, NotificationFeedSkeleton} from 
 import UserFeed, {UserFeedProps, UserFeedSkeleton} from './components/UserFeed';
 import UserProfile, {UserProfileProps, UserProfileSkeleton} from './components/UserProfile';
 import PrivateMessages, {PrivateMessagesProps, PrivateMessagesSkeleton} from './components/PrivateMessages';
+import UsersList, {UsersListProps, UsersListSkeleton} from './components/UsersList';
+import FeedObjsList, {FeedObjsListProps, FeedObjsListSkeleton} from './components/FeedObjsList';
 
 export {
   Category,
@@ -35,5 +37,11 @@ export {
   UserProfileSkeleton,
   PrivateMessages,
   PrivateMessagesProps,
-  PrivateMessagesSkeleton
+  PrivateMessagesSkeleton,
+  UsersList,
+  UsersListProps,
+  UsersListSkeleton,
+  FeedObjsList,
+  FeedObjsListProps,
+  FeedObjsListSkeleton
 };